improve blender scripts

+++ PYTHON/BLENDER/common/__init__.py
@@ -29,8 +29,19 @@ class Starter(bpy.types.Operator):
     blend = StringProperty()
 
     def execute(self, context): 
+        # Open if blend given
+        if self.blend and os.path.exists(self.blend):
+            log.yaml('Loading blend file', self.blend)
+            bpy.ops.wm.open_mainfile(filepath=self.blend)
+
         # Cycles
-        context.scene.render.engine = 'CYCLES' 
+        scene = context.scene
+        scene.render.engine = 'CYCLES' 
+        scene.cycles.samples = 64
+
+        if scene.world is None:
+            scene.world = bpy.data.worlds.new("World")
+        scene.world.cycles.sample_as_light = True
 
         def remove_object(o):
             iters = {
@@ -39,24 +50,20 @@ class Starter(bpy.types.Operator):
             }
             o_data = o.data
             o_iter = iters.get(o.type, None)
-            scene = context.scene
-            scene.objects.unlink(o)
+            if o.name in scene.objects:
+                scene.objects.unlink(o)
             bpy.data.objects.remove(o)
             if o_iter:
                 o_iter.remove(o_data)
 
-        # Open if blend given
-        if self.blend and os.path.exists(self.blend):
-            log.yaml('Loading blend file', self.blend)
-            bpy.ops.wm.open_mainfile(filepath=self.blend)
         bad_mesh = {'Cube'}
         bad_lamp = {'Lamp'}
         bad = bad_mesh | bad_lamp
-        # Remove the default cube mesh
+        # Remove unwanted objects
         for o in bpy.data.objects:
             names = {o.name, o.data.name}
             if names & bad:
-                log.yaml('Removed', o.name)
+                log.yaml('Removed', names)
                 remove_object(o)
 
         # Create the sun
@@ -64,26 +71,26 @@ class Starter(bpy.types.Operator):
             bpy.ops.object.lamp_add(**{
                 'type': 'SUN',
             })
-            sun = bpy.context.active_object
+            sun = context.active_object
             sun.data.name = 'Sun'
             sun.name = 'Sun'
 
-        area_locs = {
+        good_area = {'Area', 'Under', 'Over'}
+        area_loc = {
             'Area': (2.0, 2.0, 10.0),
             'Over': (0.0, 0.0, 1.0),
             'Under': (0.0, 0.0, -1.0),
         }
         area_size = {
-            'Area': 0.5,
+            'Area': 0.6,
             'Over': 0.5,
             'Under': 0.5,
         }
-        area_bright = {
-            'Area': 10000,
-            'Over': 100,
-            'Under': 100,
+        area_shadow = {
+            'Area': True,
+            'Over': False,
+            'Under': False,
         }
-
         # Create an area lamp
         def make_area(_name, _location):
             if _name in bpy.data.lamps:
@@ -92,28 +99,27 @@ class Starter(bpy.types.Operator):
                 'location': _location,
                 'type': 'AREA',
             })
-            area = bpy.context.active_object
+            area = context.active_object
             area.data.name = _name
             area.name = _name
 
         # Set up some area lamps
-        for name,loc in area_locs.items():
+        for name in good_area:
+            loc = area_loc[name]
             make_area(name, loc)
             area = bpy.data.lamps[name]
             area.shadow_buffer_soft = 100
             area.size = area_size[name]
-            a_bright = area_bright[name]
-            a_pow = a_bright * (loc[-1] ** 2)
-            area.energy = a_pow * (area.size ** 2)
+            # Calculate energy
+            area.energy = mover.energy(area, loc[-1])
+            area.cycles.cast_shaddow = area_shadow[name]
 
         # Set up the sun
         sun = bpy.data.lamps['Sun']
         sun.cycles.cast_shadow = False
         sun.energy = 2
 
-        #remove_lamp('Area')
-        #remove_lamp('Sun')
-
+        log.yaml('Lamps', bpy.data.lamps.keys())
         return {'FINISHED'}
 
 
@@ -127,6 +133,17 @@ class Stopper(bpy.types.Operator):
     output = StringProperty()
     movie = BoolProperty()
 
+    def device(self, context):
+        # Show available devices
+        my_sys = context.user_preferences.system
+        my_devices = my_sys.bl_rna.properties['compute_device']
+        my_types = my_sys.bl_rna.properties['compute_device_type']
+        # Show available types and devices
+        msg_d = 'Device {} from'.format(my_sys.compute_device)
+        msg_t = 'Type {} from'.format(my_sys.compute_device_type)
+        log.yaml(msg_d, [t for t in my_devices.enum_items])
+        log.yaml(msg_t, [t for t in my_types.enum_items])
+
     def render(self, context):
         scene = context.scene
         render = scene.render
@@ -137,29 +154,35 @@ class Stopper(bpy.types.Operator):
         # Path as the output of the scene
         render.filepath = self.output
         log.yaml('Rendering', self.output)
+        # Resolution double actual output
+        render.resolution_x = 900 * 2
+        render.resolution_y = 500 * 2
         # Render images to the output files
         if self.movie:
             log.yaml('Frames', frames)
-            pather.make(self.output)
             # Set frame to first frame
             scene.frame_set(frames[0])
             render.ffmpeg.codec = 'H264'
             render.ffmpeg.format = 'MPEG4'
             render.fps_base = render.fps / 24
-            render.resolution_x = 900 * 2
-            render.resolution_y = 500 * 2
             render.image_settings.file_format = 'FFMPEG'
             # Animate movie
             bpy.ops.render.render(**{
                 'animation': True
             })
             return
+        # Look at meshes
+        mover.look(scene)
+        #bpy.ops.view3d.view_orbit(angle=90.0, type='ORBITLEFT')
+        render.image_settings.file_format = 'PNG'
         bpy.ops.render.render(**{
             'write_still': True
         })
+
     def execute(self, context): 
         # Render if file given
         if self.output:
+            self.device(context)
             self.render(context)
 
         # Save if blend given
